Enemy System Part1

+++ Assets/_Game/Scripts/UI/Healthbar.cs
@@ -6,18 +6,19 @@ public class Healthbar : MonoBehaviour
     public Gradient gradient;
     public Image fill;
 
-    public void SetmaxHealth(int Health)
-    {  Slider.maxValue = Health;
-       Slider.value = Health;
+    public void SetmaxHealth(float Health)
+    {
+        Slider.maxValue = Health;
+        Slider.value = Health;
 
-        fill.color = gradient.Evaluate(1f) ;
+        fill.color = gradient.Evaluate(1f);
     }
-    
-    public void SetHealth(int health)
+
+    public void SetHealth(float health)
     {
         Slider.value = health;
         fill.color = gradient.Evaluate(Slider.normalizedValue);
     }
-    
+
 
 }
